refactor: update landing page components and create nav component
