+++ src/pages/about/AboutPage.jsx
@@ -1,4 +1,5 @@
 import React from "react";
+import "./AboutPage.scss";
 
 export default function AboutPage() {
   return (
