+++ src/pages/contact/ContactPage.jsx
@@ -1,4 +1,5 @@
 import React from "react";
+import "./ContactPage.scss";
 
 export default function ContactPage() {
   return (
