+++ src/pages/plansAndPricing/PlansAndPricingPage.jsx
@@ -1,9 +1,10 @@
 import React from "react";
+import "./PlansAndPricingPage.scss";
 
 export default function PlansAndPricingPage() {
   return (
     <div>
-      <h1>Plans And Priciing</h1>
+      <h1>Plans And Pricing</h1>
     </div>
   );
 }
