+++ src/routes.jsx
@@ -4,9 +4,9 @@ import Layout from "./Layout";
 import LandingPage from "./pages/landing/LandingPage";
 import ErrorPage from "./pages/error/ErrorPage";
 import AboutPage from "./pages/about/AboutPage";
-import ContactPage from "./pages/about/ContactPage";
-import FeaturesPage from "./pages/about/FeaturesPage";
-import PlansAndPricingPage from "./pages/about/PlansAndPricing";
+import ContactPage from "./pages/contact/ContactPage";
+import FeaturesPage from "./pages/features/FeaturesPage";
+import PlansAndPricingPage from "./pages/plansAndPricing/PlansAndPricingPage";
 
 export default createBrowserRouter([
   {
